[4301] Rename ResponseConverterFunctionSelector to ResponseConverterFunctionUtil
